FallbackExceptionMapper logs stack traces (CORE-288)
The FallbackExceptionMapper exists to capture exceptions that aren't
otherwise handled by a JAX-RS application and map them into nice JSON
responses.  However, in some cases where the error message is very
succint, or generic, it doesn't help developers pinpoint where that
error actually occurred in their code so they can debug and fix it.

The mapper is modified to now explicitly log the stack trace for the
error at ERROR level in the logs to aid developers in diagnosing the
root cause of such errors.