/healthz caches health status (CORE-422)
This modifies the AbstractHealthResource to cache the applications
computed health status for a brief period (30 seconds).  This provides a
DoS mitigation in case of the applications health check involving
expensive computation, or probes to the services it depends on, that
would otherwise provide a DoS vector for a malicious user to attack the
underlying services.

Also modifies AbstractApplication to developers need to be more explicit
about choosing to implement, or not implement, a /healthz endpoint in
their applications.