Improved KafkaSink send() behaviour
Noticed when trying to use KafkaSink outside of test cases that in some
scenarios it was silently swallowing errors produced by Kafka so caller
could think data was going to Kafka when it wasn't.  Improved
KafkaSink.send() implementation to capture and surface KafkaProducer's
async errors on subsequent send()/close() calls.

Added various new unit and integration tests around this, and updated
some existing negative tests that relied on the old buggy behaviour to
be aware that errors are now thrown in those cases.

Documentation is updated to reflect the new behaviour and clarify how
and when errors are surfaced.

Also fixes a related bug with using Projector commands in
unit/integration test scenarios that could leave a projection running
indefinitely.