Improved error handling around Live Reporter shutdown
Improves the error handling in LiveReporter and LiveErrorReporter so
that in the event of errors occurring while trying to heartbeat/report
errors the reporter does not abruptly terminate.

Also when these reporters are torn down better handle any errors that
can occur in the shutdown which would otherwise hide/confuse the real
shutdown cause.

Added additional unit tests across all these areas and renamed existing
tests to be more meaningful and adopt Given, When, Then style