Test coverage for RejectSink
Adds test coverage for RejectSink, including cleaning up the existing
tests for various basic sinks into their own test classes with some
common functionality in an abstract base class.  This makes it easier to
find tests, and tests are renamed with more descriptive names and
refactored into Given, When, Then style.

Further refactored the builder for RejectSink and FilterSink to make it
more natural to use.

Added documentation for the new RejectSink