Support serializing Problem as plain text
This fixes a bug that has cropped up a few times as new services are
built out that if an API is designed to support simpler text/plain
responses but an error occurs our ExceptionMapper's turn it into a
Problem response which can't be serialized as text/plain.

To resolve this we add a new ProblemPlainTextWriter and register it in
our AbstractApplication so serializing Problem responses to text/plain
is supposed out of the box.

A related change is that Problem.toResponse() no longer sets the
Content-Type always to application/problem+json, instead it allows the
JAX-RS framework to set that correctly based on the HTTP request content
negotation it has already done.