Improved Problem Response Content-Type headers
Iterates on a change made in the 0.25.0 release where we stopped setting
the Content-Type header at all to allow the JAX-RS server to do content
negotation properly.  However, turns out that in some corner cases this
results in a Problem response being defaulted to
application/octet-stream responses which is rarely useful/helpful.

Therefore this improves the behaviour to make a best effort to add a
recognised Content-Type header, when it is compatible with the
acceptable content types indicated by the clients Accept header. A new
toResponse(HttpHeaders) overload is added to facilitate this, and
existing callers of toResponse() updated to call the new method instead.

Also greatly expands the test coverage in this area to validate various
scenarios around different combinations of Accept header, resource
@Produces headers and the resulting Content-Type header.