Improved KafkaEventSource.resetOffsets() (CORE-684)
Improved the implementation of the resetOffsets() method to account for
the fact that like committing offsets this can only be done from the
polling thread.  If called from another thread the offset resets are
recorded and applied next time the polling thread is preparing to poll()
the KafkaConsumer.  Lots more test coverage added around these corner
cases.