Support for testing external commands (CORE-705)
Existing command testing focuses on running the command in the same JVM
as the test.  This is useful in most cases but it doesn't catch
dependency scoping errors that can creep in when something is placed
into test scope unintentionally.

This commit adds new support to SmartCacheCommandTester for running
tests where the command is spawned as it would be in a real deployment.
This allows for easily writing test cases that can verify the real
application will run and have all the relevant dependencies available to
actually do its job correctly in a real deployment.