Introduce EventHeader interface (CORE-761)
This provides for the ability to access the raw byte[] value of a
header, and also a new RawHeader implementation for event sources like
Kafka that don't want to convert header values to String unnecessarily.

Various APIs are updated to use the EventHeader interface in preference
to the existing Header implementation class.