Allow ProjectorDriver logging to be labelled (CORE-905)
When multiple ProjectorDriver's are running in an application there
logging is interleaved which can make it hard to parse/debug.  This
commit adds the ability to specify a log label on a driver which is then
inserted at the start of all logging from the ProjectorDriver to make
the logs easier to read.