Fix RdfPayload Kafka serialization edge case (CORE-978)
Where a RdfPayload was deserialized from an event declaring a
Content-Type for an RDF serialization that does not support named graphs
attempting to serialize it back to Kafka would result in a "No dataset
writer" error from Jena.  Now in this scenario we can successfully
serialize the payload back to Kafka provided it only contains a default
graph.

In the event that we are trying to serialize a RdfPayload with a graph
Content-Type header and named graphs are present a more informative
error is now generated indicating that the Content-Type header is wrong.