Improved Auth Server integration (CORE-961)
Discovered during integration testing that the UserInfoFilter in Core
didn't integrate with the new AttributesStoreAuthServer in RDF-ABAC
meaning when using the new Auth Server with Core apps like SC-Search
users had no access to data as they had no attributes.

This commit applies the approach already taken in SC-Graph (which
doesn't use JAX-RS) of having the UserInfoFilter populate the
AttributeStoreAuthServer caches as requests are processed allowing the
application to correctly access user attributes.