Improved JAX-RS error handling (CORE-1092)
In some bad request scenarios a very large stack trace could get dumped
due to Jersery producing a MultiException which contains many separate
Exceptions.  We now have a specific ExceptionMapper for these which
avoids logging stack traces and just reports general details to the end
user, while logging the useful details in the server logs.

As part of this change the FallbackExceptionMapper and
FailureLoggingFilter were also improved to include additional details in
the logging to make it easier to determine what provoked the error