Add fallback error handler
As noticed while testing updated UIs in dev clusters if the UI sends a
malformed Content-Type header then request processing never reaches
JAX-RS handling, and thus bypasses our normal error handling and logging
logic.

To rectify this we now install a custom ErrorPageGenerator so that in
the event of one of these class of errors we now properly log the error,
and provide a cleaner HTML error page in our response.